Probs

Class to represent probabilities used in mutation and crossover operations.

Details

This class provides a convenient way to manage and update probabilities used in mutation and crossover operations. It allows you to set individual probabilities for various operations, such as overall mutation, overall crossover, parameter mutation, parameter crossover, and group structure mutation. Future more flexible classes are likely to inherit from this base class.

Public fields

p_overall_mutate

(numeric(1))
Numeric value between 0 and 1 specifying the overall mutation probability.

p_overall_crossover

(numeric(1))
Numeric value between 0 and 1 specifying the overall crossover probability.

p_param_mutate

(numeric(1))
Numeric value between 0 and 1 specifying the hyperparameter mutation probability.

p_param_crossover

(numeric(1))
Numeric value between 0 and 1 specifying the hyperparameter crossover probability.

p_groupstructure_mutate

(numeric(1))
Numeric value between 0 and 1 specifying the group structure mutation probability.

p_groupstructure_crossover

(numeric(1))
Numeric value between 0 and 1 specifying the group structure crossover probability.

Methods

Public methods

Probs$new()

Probs$update()

Probs$clone()

Method new()

Creates a new instance of this R6 class.

Usage

Probs$new(
  p_overall_mutate = 0.3,
  p_overall_crossover = 0.7,
  p_param_mutate = 0.2,
  p_param_crossover = 0.5,
  p_groupstructure_mutate = 0.2
)

Arguments

p_overall_mutate

(numeric(1))
Numeric value between 0 and 1 specifying the overall mutation probability. The default value is 0.3.

p_overall_crossover

(numeric(1))
Numeric value between 0 and 1 specifying the overall crossover probability. This overall crossover probability is also used for the group structure crossover probability. The default value is 0.7.

p_param_mutate

(numeric(1))
Numeric value between 0 and 1 specifying the hyperparameter mutation probability. The default value is 0.2.

p_param_crossover

(numeric(1))
Numeric value between 0 and 1 specifying the hyperparameter crossover probability. The default value is 0.5.

p_groupstructure_mutate

(numeric(1))
Numeric value between 0 and 1 specifying the group structure mutation probability. The default value is 0.2.

Method update()

Updates the probabilities.

Currently has no effect because probabilities are treated as non-adaptive.

Usage

Probs$update()

Returns

Invisible (NULL)

Method clone()

The objects of this class are cloneable with this method.

Usage

Probs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.