Selected Features Proxy

Proxy measure that simply returns 0. Can be used as a proxy measure which is updated within an optimizer. Only to be used internally.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

mlr_measures$get("selected_features_proxy")
msr("selected_features_proxy")

Meta Information

Task type: “NA”

Range: \([0, \infty)\)

Minimize: TRUE

Average: macro

Required Prediction: “response”

Required Packages: mlr3

Parameters

Empty ParamSet

Super class

mlr3::Measure -> MeasureSelectedFeaturesProxy

Methods

Public methods

MeasureSelectedFeaturesProxy$new()

MeasureSelectedFeaturesProxy$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureSelectedFeaturesProxy$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureSelectedFeaturesProxy$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.