Monotonicity Detector

Detects whether a feature should have a monotone increasing or decreasing effect (if enforced) and also quantifies the tendency of a monotone relationship between the feature and the target.

Only works with integer or numeric features. Logical features must be converted to integers.

Public fields

task

(mlr3::Task)
Task.

classification

(logical(1))
Whether the task is a classification task.

data

(data.table::data.table)
data.table of data of the mlr3::Task.

n_features

(integer(1))
Number of features part of the mlr3::Task.

feature_names

(character())
The names of the features of the mlr3::Task.

feature_types

(character())
The types of the features of the mlr3::Task.

y_name

(character(1))
The names of the target of the mlr3::Task.

rho_table

(data.table::data.table)
data.table with as many rows as features with the feature_name column indicating the feature and the rho column indicating Spearman's Rho.

unconstrained_weight_table

(data.table::data.table)
data.table with as many rows as features with the feature_name column indicating the feature and the unconstrained_weight column indicating the absolute Spearman's Rho scaled rescaled to [0.2, 0.8].

Methods

Public methods

MonotonicityDetector$new()

MonotonicityDetector$compute()

MonotonicityDetector$get_sign()

MonotonicityDetector$compute_unconstrained_weights()

MonotonicityDetector$clone()

Method new()

Creates a new instance of this R6 class.

Usage

MonotonicityDetector$new(task)

Arguments

task

([mlr3::Task)
Task.

Method compute()

Fits a decision tree for each feature and the target variable using subsampled data and calculates Spearman's Rho between the prediction (numeric for regression and probability of the positive class for binary classification) and the feature. Repeats this process for each feature ten times and averages the Spearman's Rhos to fill the $rho_table Subsampling fraction is set to 0.9 and the decision tree is a standard mlr3::LearnerRegr or mlr3::LearnerClassif. FIXME: expose hyperparameters etc.

Usage

MonotonicityDetector$compute()

Returns

Invisible (NULL)

Method get_sign()

Method to determine the sign of a given feature based on the average Spearman's Rho correlation if it should be constrained to have a monotone effect.

Usage

MonotonicityDetector$get_sign(feature_name)

Arguments

feature_name

(character(1))
Feature for which the sign should be determined

Returns

(integer()) vector indicating the sign of the feature if it should be constrained to have a monotone effect

Method compute_unconstrained_weights()

Method to construct the unconstrained weights table from the average Spearman's Rhos by taking the absolute value and scaling it to [0.2, 0.8].

Usage

MonotonicityDetector$compute_unconstrained_weights()

Returns

Invisible (NULL)

Method clone()

The objects of this class are cloneable with this method.

Usage

MonotonicityDetector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.